irgen(refactor): move context to azuki.rs
fix in main
